Temporary fix to allow one modification to a field to be summarized correctly.
This does not fix the more serious problem that fields in WALA IR
dont have SSA form. Therefore, when a field is written into more than
once inside a region, it essentially makes that branch side of the region
infeasible because the same symbolic variable corresponding to the field
output has to now equal the two written values simultaneously.