Making the instantiateRegionOutput in VeritestingListener be a stack instead of being a field.
This fix is required because instantiateRegionOutput can get overwritten
with another object of InstantiateRegionOutput on an SPFCase choice later
causing us to lose track of it in the current stack frame of
VeritestingListener.executeInstruction.

In another world, executionInstruction would have called itself
tail-recursively causing instantiateRegionOutput to be stack frame-specific
instead of us having to maintain a stack for it.